VBID-7212 Vistar project should be able to serve JS projects
Vistar project successfully serving an example project. To put context to the
code in `//vistar/ui/modstyle/test`, the compiled app is available here:
https://s3.amazonaws.com/mark.vistarmedia.com/js-demo/index.html

There were some small fixes to the typescript rules and js-squish'er to get
things running in the end, but the bulk of the cod here is in the test app
itself. See the README at //vistar/ui/modstyle/README.md for a good synopsis of
what's going on.

Change-Id: I78097f94339be018c0b973388d7964cf123524ed