VBID-7269 Make Bazel typescript tests play nice with Mocha
When initializing TS repositories, keep a reference to mocha's typedef file
from npm. Pass that in as a source file to tsc when transpiling a test to JS.

I'm working on VBID-7217 which involves upgrading the Bazel version in
cortex-apps. I started looking at this because if I didn't fix it, I would have
to touch every test file in that directory

Change-Id: I51649c43ffabd6243039ad49d7f754c114eba83a